feat(kunlunxin): batch3 kunlunxin fixes on upstream master baseline
Batch 3 verification closure (303 ops): 227 files (149 modified + 78 added)
on top of upstream/master fcdd6ee97.
- 163 PASS / 129 BLOCKED / 11 SKIP (dtype-equal-weight Gems Speedup >= 0.8 gate)
- Pure vendor-dir changes under _kunlunxin/ plus tests/benchmark updates
- No CPU/ATen/native fallback used as fix or PASS basis

+++ src/flag_gems/runtime/backend/_kunlunxin/ops/dequantize.py
@@ -51,7 +51,11 @@ def _dequantize_kernel(
     # valid lanes. Lanes past n_elements read undefined bytes and land in
     # the padded tail of `out_ptr`, which the caller never exposes.
     raw = tl.load(x_ptr + offsets, mask=offsets < n_elements)
-    values = raw.to(tl.float32)
+    # i8 -> f32 via i32: the direct `arith.sitofp` on a vector of 64 x i8 is
+    # mis-lowered by TritonXPU's ConvertTritonXPUToLLVM ("size mismatch when
+    # packing elements for LLVM struct expected 8 but got 2"), so the sign
+    # extension is done in two 1:1 steps (i8 -> i32 -> f32).
+    values = raw.to(tl.int32).to(tl.float32)
     if UNSIGNED:
         values = tl.where(values < 0.0, values + 256.0, values)
     tl.store(out_ptr + offsets, (values - zero_point) * scale)
